Split ORCID tests in a separate file
